Patch from Frank Rast: Solves a memory leak caused by the member variable ownsPool never being initialized and thus containing random data.
Additionally it solves an issue with the copy and assignment of Pool and SubPool. The argument must always be (const T&) and there should not be any
function body in order to avoid accidental use or assignment of this class. The linker will not give an error because nothing refers to copy
or assignment.

* \src2\CONTRIBUTERS
	Added Frank Rast.

* \src2\NSvn.Core\Pool.h
	(.ctor * 2) - Initialize ownsPool.
	(copy .ctor and operator=) - Take in a const Pool&, remove the implementation.
	(SubPool :: copy .ctor and operator=) - Ditto.


git-svn-id: https://ctf.open.collab.net/svn/repos/ankhsvn/trunk@2540 ecf7f2a6-7b29-9143-b1c6-df2d17ec8657