Added AnkhSvn and AnkhSvn.Ids projects for both VS2005 and VS2008.
AnkhSvn is a very minimal VS package containing nothing more than a few 
command definitions. All command definitions are compiled to a ctc file
which is embedded in the managed resources. (MS framework).

In the Gui subdirectory I added a conversion script from a much friendlier
xml file format to define the UI structure, which reads the definitions 
in  then new AnkhSvn.Ids assembly to get the command values used within 
the code. (This part abstracts all the VS menu and toolbar integration 
from the csharp code.)

The .csproj file is completely altered to work with both the VS2005 
and VS2008 SDK. It's dependency assemblies change between the two SDK 
releases. To make sure you don't accidently build code which is not
compatible with the other SDK, most code must not be embedded directly
within this assembly.

* AnkhSvn
* AnkhSvn/Guids.cs
* AnkhSvn/AnkhSvn.snk
* AnkhSvn/GlobalSuppressions.cs
* AnkhSvn/Key.snk
* AnkhSvn/Gui
* AnkhSvn/Gui/Gui-To-Ctc.xsl
* AnkhSvn/Gui/Gui.xsd
* AnkhSvn/Gui/Gui.xsx
* AnkhSvn/Gui/AnkhSvn.xml
* AnkhSvn/VSPackage.resx
* AnkhSvn/AnkhSvn.vsct
* AnkhSvn/AnkhSvn.csproj
* AnkhSvn/AnkhSvnPackage.cs
* AnkhSvn/Resources
* AnkhSvn/Resources/Package.ico
* AnkhSvn/Resources/Images_32bit.bmp
* AnkhSvn/Resources.resx
* AnkhSvn/Properties
* AnkhSvn/Properties/AssemblyInfo.cs
* AnkhSvn/Resources.Designer.cs
  See above

* AnkhSvn/Gui/ctc
  Directory containing the generated ctc file. You will probably see
  errors refering to files here if you mis-edit the AnkhSvn.xml file

* AnkhSvn.Ids
* AnkhSvn.Ids/AnkhSvn.Ids.snk
* AnkhSvn.Ids/AnkhCommandGroup.cs
* AnkhSvn.Ids/AnkhId.cs
* AnkhSvn.Ids/AnkhCommand.cs
* AnkhSvn.Ids/AnkhSvn.Ids.csproj
* AnkhSvn.Ids/AnkhCommandMenu.cs
* AnkhSvn.Ids/Properties
* AnkhSvn.Ids/Properties/AssemblyInfo.cs
  Added assembly containing a few enum and constant classes. These
  values are read by the Gui-To-Ctc convertion script and embedded
  in the compiled resources. The same values are send back from VS
  when updating. This assembly stays in memory within VS after 
  updating. (VS does not lock it)


git-svn-id: https://ctf.open.collab.net/svn/repos/ankhsvn/trunk@3363 ecf7f2a6-7b29-9143-b1c6-df2d17ec8657