Forward the new SvnActions to the SccProvider.
The SccProvider will then (on the next idle processing) check if there are
any websites and if there are pass the add and removes to the SccProjectData.

The SccProjectData in its turn will look for added or removed files and
push this information into websites. (The only well-known project type that
doesn't look at a project file to determine what is inside a project or not).

* src/Ankh.Scc/Ankh.Scc.csproj
  Add file
* src/Ankh.Scc/AnkhSccProvider.Projects.cs
  Cache commandservice instance... and push idle command on scc refresh actions
* src/Ankh.Scc/ProjectMap/SccProjectData.ProjectRefresh.cs
  New file to handle adding and removing files from website projects.
  (Might be moved to a separate service in future cleanup)
* src/Ankh.Scc/ProjectNotifier.cs
  Forward a subset of the SvnActions to the Scc provider.


git-svn-id: https://ctf.open.collab.net/svn/repos/ankhsvn/trunk@6604 ecf7f2a6-7b29-9143-b1c6-df2d17ec8657